requirements.txt
openai==1.3.0
python-dotenv==1.0.0
pygame==2.5.0